+++ src/test/java/org/elasticsearch/test/store/MockDirectoryHelper.java
@@ -36,6 +36,7 @@ import org.elasticsearch.index.store.fs.MmapFsDirectoryService;
 import org.elasticsearch.index.store.fs.NioFsDirectoryService;
 import org.elasticsearch.index.store.fs.SimpleFsDirectoryService;
 import org.elasticsearch.index.store.ram.RamDirectoryService;
+import org.elasticsearch.test.TestCluster;
 
 import java.io.IOException;
 import java.util.Random;
